# stop overwriting zend_execute_ex; start using observer api