PHP 8 introduces an observer API, which allows better monitoring and observability of PHP applications without causing significant performance issues or unexpected behavior. Datadog has written about it in [this article](https://href.li/?https://www.datadoghq.com/blog/engineering/php-8-observability-baked-right-in/). Newrelic’s php-agent [has implemented the new api](https://href.li/?https://github.com/newrelic/newrelic-php-agent/blob/4f470dac6640429416adff5b3937fc36c36f1f96/agent/php_execute.c#L2139-L2174).

- Avoids the limited native C stack that could lead to stack overflow and crashes when `zend_execute_ex` intercepted all PHP function calls
- Allows targeting specific PHP functions of interest rather than intercepting every single PHP function call, improving performance
- Works compatibly alongside the new JIT compiler introduced in PHP 8
- Handles forwarding hooks between extensions to avoid "noisy neighbor" issues and crashes from improper forwarding
- Provides a standard "observer" concept and API for future expansion of observability in the PHP engine