@mreishus 2024-05-10
Benchmarked the performance impact of tombs using the Observer API compared to the previous zend_execute_ex approach and a baseline without tombs.

Methodology:
- Tested two pages: Homepage and a Single Post page (WordPress.org installed locally using php-fpm and nginx)
- OpCache enabled
- Used a curl script to measure time-to-first-byte (TTFB) over 250 requests
- Compiled with -O2 optimization
- Verified correct number of tombs were created in each scenario

**Homepage:**
| Scenario                | p50 (ms) | p75 (ms) |
|-------------------------|----------|----------|
| No Tombs                | 43.5     | 44.3     |
| Tombs - zend_execute_ex | 44.4     | 45.1     |
| Tombs - Observer API    | 45.6     | 46.7     |

**Single Post:**
| Scenario                | p50 (ms) | p75 (ms) |
|-------------------------|----------|----------|
| No Tombs                | 37.4     | 38.1     |
| Tombs - zend_execute_ex | 38.7     | 40.7     |
| Tombs - Observer API    | 39.1     | 39.9     |

Results show that switching to the Observer API has a minor impact on TTFB compared to the zend_execute_ex approach:

For the Homepage:
- Median TTFB increased by ~1.2ms (from 44.4ms to 45.6ms) compared to zend_execute_ex
- 75th percentile TTFB increased by ~1.6ms (from 45.1ms to 46.7ms)

For the Single Post:
- Median TTFB increased by ~0.4ms (from 38.7ms to 39.1ms) compared to zend_execute_ex
- 75th percentile TTFB increased by ~0.8ms (from 40.7ms to 39.9ms)

Overall, using the Observer API provides better compatibility with PHP 8 while still maintaining reasonable performance. The minor overhead seems acceptable given the benefits of using the new API.